Add comprehensive unit, integration, and E2E test suites
Replace the bash smoke test with 421 xUnit v3 tests across three
projects, all runnable via plain `dotnet test` and wired into CI:

- WaveBox.Core.Tests (185, parallel): extensions, User crypto, Utility,
  SQLiteConnectionPool, and reflection tests asserting every ORM model
  and Subsonic DTO is rooted in ModelTypeRegistry/SubsonicDtoRegistry,
  catching NativeAOT rooting drift at test time.
- WaveBox.Server.Tests (217): unit layer (UriWrapper, SubsonicMapper,
  XML serializer, transcoders, image resize) plus a serialized
  integration collection against a real temp SQLite database: schema
  migration idempotency, repositories, playlist logic, a real TagLib
  folder scan of a generated MP3 fixture, ServerSettings JSONC
  round-trips, and SubsonicAuth error codes.
- WaveBox.E2E.Tests (19): full port of Scripts/smoke-test.sh. A fixture
  boots the real server binary on a random free port with an isolated
  root; WAVEBOX_E2E_BINARY selects the NativeAOT publish output in CI,
  falling back to the local build for development.
- WaveBox.TestFixtures: shared MP3 generator (C# port of
  make_fixture.py, dropping the python3 dependency), temp-root,
  free-port, and conf-seeding helpers.

Testability changes: WAVEBOX_ROOT/WAVEBOX_TEMP env overrides for
ServerUtility.RootPath()/ServerInfo.TempFolder (also enables Windows
E2E isolation), InternalsVisibleTo for the test projects, and three
private pure helpers widened to internal.

Bug fixes, each with a regression test:
- IntExtensions.ToTimeString dropped the minutes segment at H>0, M=0
  (3600s rendered "1:00" instead of "1:00:00")
- AbstractTranscoder.GetHashCode inverted null ternary NRE'd exactly
  when Item was null
- SQLiteConnectionPool.CloseAllConnections never re-enabled pooling
  after a backup
- Utility.RandomString used a shared non-thread-safe Random to generate
  session IDs; now Random.Shared
- Config "port" was Int16, so values above 32767 failed JSON parsing
  and silently fell back to 6500; now int end to end
- StringExtensions.RemoveByteOrderMark used culture-sensitive
  StartsWith, stripping the first character of every string under ICU;
  now ordinal

CI: fix the stale push trigger (master -> main), add a fast 3-OS test
job gating the AOT publish matrix, add NuGet caching, and replace the
smoke-test step with the E2E suite against the published binary,
now including win-x64. Scripts/smoke-test.sh and
tests/fixtures/make_fixture.py are retired.

+++ WaveBox.Core/src/Injection/Interfaces/IServerSettings.cs
@@ -11,7 +11,7 @@ namespace WaveBox.Core {
 
         bool PrettyJson { get; }
 
-        short Port { get; }
+        int Port { get; }
 
         string Theme { get; }
